Separate out the time slicer from the rest of reco
